integration_package/requirements.txt
# Core data processing
pandas==2.2.0
openpyxl==3.1.2  # Excel file handling
numpy==1.26.3

# Google API clients
googlemaps==4.10.0
google-cloud-storage==2.14.0

# Web scraping (for zoning data)
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.17.2

# Database
sqlalchemy==2.0.25
sqlite3  # Built-in

# Interactive dashboard
folium==0.15.1  # Interactive maps
plotly==5.18.0  # Interactive charts
streamlit==1.30.0  # Web dashboard
streamlit-folium==0.17.4

# Mail merge
python-docx==1.1.0  # Word documents
jinja2==3.1.3  # Template engine

# Utilities
python-dotenv==1.0.0  # Environment variables
tqdm==4.66.1  # Progress bars
geopy==2.4.1  # Geocoding utilities
fuzzywuzzy==0.18.0  # String matching for deduplication
python-Levenshtein==0.23.0  # Speed up fuzzy matching

# Data validation
pydantic==2.5.3
email-validator==2.1.0

# Development
jupyter==1.0.0  # For data exploration
ipykernel==6.28.0
